refactor(web): change register technology button title
