+++ packages/web/components/LandingPage/Header/index.js
@@ -1,4 +1,4 @@
 export { default as Header } from './Header';
 export { default as UserHeader } from './UserHeader';
 export { default as LanguageSwitcher } from './LanguageSwitcher';
-export { default as NewTechnologyButton } from './NewTechnologyButton';
+export { default as NewSolutionButton } from './NewSolutionButton';
