+++ packages/web/public/static/locales/en/common.json
@@ -29,7 +29,7 @@
   "recentSolutions": "<span>Recent</span> Solutions",
   "region": "Region",
   "register": "Register",
-  "registerTechonology": "<span>Register your</span> technology",
+  "registerSolution": "<span>Register your</span> solution",
   "registerTitle": "<span>You are about to <br /> discover a <br /> world of opportunities</span>",
   "registerYourUser": "REGISTER YOUR USER",
   "relatedSolutions": "Related Solutions",
