+++ packages/web/public/static/locales/pt/common.json
@@ -29,7 +29,7 @@
   "recentSolutions": "<span>Soluções</span> Recentes",
   "region": "Região",
   "register": "Cadastrar",
-  "registerTechonology": "<span>Cadastre sua</span>tecnologia",
+  "registerSolution": "<span>Cadastre sua</span>solução",
   "registerTitle": "<span>Você está prestes <br /> a descobrir um<br /> mundo de oportunidades</span>",
   "registerYourUser": "CADASTRE SEU USUÁRIO",
   "relatedSolutions": "Soluções relacionadas",
